feat: Support default query options
Query options enables a backend system to define custom query options. Query options were not provided with default values. As a result each new query was initialized with empty options which makes the behavior of the system unclear for the user. This change makes it possible to define default options in the backend. In addition to this the backend query has been enhanced with the current selected query options. This makes it possible for the backend to tailor the query options based on the current selection.